firewalldb+refactor: move kvdb priv map code to dedicated file
In preparation for adding a file dedicated to the SQL implementation of
the privacy mapper DB.
